[add] CLI output based on CommanderJSX [add] Dependency Badge based on Libraries.io [optimize] update Upstream packages
