+++ .npmignore
@@ -1,0 +1,3 @@
+.vscode/
+.husky/
+.github/
