+++ ReadMe.md
@@ -2,9 +2,10 @@
 
 A **file downloader** based on [Git][1] CLI
 
-[![CI & CD](https://github.com/idea2app/get-git-folder/actions/workflows/main.yml/badge.svg)][2]
+[![NPM Dependency](https://img.shields.io/librariesio/github/idea2app/get-git-folder.svg)][2]
+[![CI & CD](https://github.com/idea2app/get-git-folder/actions/workflows/main.yml/badge.svg)][3]
 
-[![NPM](https://nodei.co/npm/get-git-folder.png?downloads=true&downloadRank=true&stars=true)][3]
+[![NPM](https://nodei.co/npm/get-git-folder.png?downloads=true&downloadRank=true&stars=true)][4]
 
 ## Usage
 
@@ -15,5 +16,6 @@ get-git-folder https://github.com/your-org/your-repo main your-folder
 ```
 
 [1]: https://git-scm.com/
-[2]: https://github.com/idea2app/get-git-folder/actions/workflows/main.yml
-[3]: https://nodei.co/npm/get-git-folder/
+[2]: https://libraries.io/npm/get-git-folder
+[3]: https://github.com/idea2app/get-git-folder/actions/workflows/main.yml
+[4]: https://nodei.co/npm/get-git-folder/
